backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic-settings
pydantic
pytest
httpx
